feat: preview share
