--- src/pages/setup/process/xiao_esp32s3/Preview.vue
@@ -1,56 +1,0 @@
-<script setup lang="ts">
-  import { watch, onMounted, onBeforeUnmount, ref } from 'vue';
-  import { Message } from '@arco-design/web-vue';
-  import { useI18n } from 'vue-i18n';
-  import useDeviceManager from '@/hooks/deviceManager';
-  import { useDeviceStore } from '@/store';
-  import { DeviceStatus } from '@/sscma';
-  import Preview from '../components/Preview.vue';
-
-  const { device } = useDeviceManager();
-  const deviceStore = useDeviceStore();
-  const invoke = ref<boolean>(false);
-  const disable = ref<boolean>(true);
-  const { t } = useI18n();
-  const previewRef = ref<InstanceType<typeof Preview> | null>(null);
-  const eventName = 'INVOKE';
-
-  const handelRefresh = async () => {
-    if (deviceStore.deviceStatus === DeviceStatus.SerialConnected) {
-      disable.value = false;
-      const isInvoke = await device.value?.isInvoke();
-      if (isInvoke) {
-        invoke.value = true;
-        deviceStore.setIsInvoke(true);
-      } else {
-        const result = await device.value?.invoke(-1);
-        if (result) {
-          invoke.value = true;
-        } else {
-          Message.error(t('workplace.preview.message.invoke.failed'));
-          invoke.value = false;
-        }
-      }
-    } else {
-      deviceStore.setIsInvoke(false);
-      disable.value = true;
-      invoke.value = false;
-    }
-  };
-  watch(() => deviceStore.deviceStatus, handelRefresh);
-
-  onMounted(async () => {
-    if (previewRef.value?.onInvoke) {
-      device.value?.addEventListener(eventName, previewRef.value?.onInvoke);
-    }
-    handelRefresh();
-  });
-
-  onBeforeUnmount(() => {
-    device.value?.removeEventListener(eventName);
-  });
-</script>
-
-<template>
-  <Preview ref="previewRef" />
-</template>
